further work on labextension
